refactor: create models
