+++ src/models/Board.ts
@@ -1,0 +1,9 @@
+export interface INewBoard {
+  title: string;
+  owner: string;
+  users: string[];
+}
+
+export interface IBoard extends INewBoard {
+  _id: string;
+}
