+++ src/models/ErrorResponse.ts
@@ -1,0 +1,4 @@
+export interface IErrorResponse {
+  status: number;
+  message: string;
+}
